Mostly optimizations. (#1152)
* Created a simplified surface data structure used for building placement and pathfinding.
AI pnly loads planet building data when needed.
Added maximum message history limit of 1000.
Various bugfixes.
Various minor optimizations.

* Reverted on-demand loading of AI building and planet surface data.

+++ src/hu/openig/model/Parameters.java
@@ -54,7 +54,7 @@ public class Parameters {
     protected int costToHitpoints = 10;
     /** What is considered a planet-nearby. */
     @LoadField
-    protected double nearbyDistance = 10;
+    protected int nearbyDistance = 10;
     /** The ratio between the fleet movement timer and speed() minute simulation time. */
     @LoadField
     protected int simulationRatio = 10;
@@ -158,7 +158,7 @@ public class Parameters {
      * What is considered a planet-nearby.
      * @return the distance
      */
-    public double nearbyDistance() {
+    public int nearbyDistance() {
         return nearbyDistance;
     }
     /**
